requirements.txt
absl-py==2.0.0
array-record==0.4.1
asgiref==3.7.2
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
bing==0.31
blinker==1.7.0
bokeh==2.4.3
bulk==0.2.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Django==4.2.7
dm-tree==0.1.8
downloader==0.98
etils==1.5.2
exceptiongroup==1.2.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2023.10.0
gast==0.4.0
google-auth==2.24.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
idna==2.10
image==1.5.33
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
namex==0.0.7
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==1.5.3
Pillow==9.0.1
promise==2.3
protobuf==3.19.6
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.25.1
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
selenium==3.141.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-datasets==4.9.2
tensorflow-estimator==2.10.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.14.0
termcolor==2.4.0
threadpoolctl==3.2.0
toml==0.10.2
tornado==6.4
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.18
wasabi==1.1.2
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
zipp==3.17.0